backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
redis==5.0.1
django-redis==5.4.0
celery==5.3.4
Pillow==10.2.0
django-filter==23.5
python-decouple==3.8
gunicorn==21.2.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
stripe==7.9.0
django-storages==1.14.2
boto3==1.34.18
openai==1.12.0
requests==2.31.0
